revamped to not modify state. test passing
