+++ integ/checkpoint_failure_test.go
@@ -2,6 +2,7 @@ package integ
 
 import (
 	"fmt"
+	v12 "k8s.io/api/core/v1"
 	"time"
 
 	"github.com/lyft/flinkk8soperator/pkg/apis/app/v1beta1"
@@ -66,7 +67,6 @@ func failingJobTest(s *IntegSuite, c *C, testName string, causeFailure func()) {
 // Tests that we correctly handle updating a job with task failures
 func (s *IntegSuite) TestJobWithTaskFailures(c *C) {
 	log.Info("Starting test TestJobWithTaskFailures")
-	c.Skip("local")
 
 	failingJobTest(s, c, "taskfailure", func() {
 		err := s.Util.ExecuteCommand("minikube", "ssh", "touch /tmp/checkpoints/fail && chmod 0644 /tmp/checkpoints/fail")
@@ -78,7 +78,6 @@ func (s *IntegSuite) TestJobWithTaskFailures(c *C) {
 // Tests that we correctly handle updating a job with a checkpoint timeout
 func (s *IntegSuite) TestCheckpointTimeout(c *C) {
 	log.Info("Starting test TestCheckpointTimeout")
-	c.Skip("local")
 
 	failingJobTest(s, c, "checkpointtimeout", func() {
 		// cause checkpoints to take 120 seconds
@@ -86,4 +85,63 @@ func (s *IntegSuite) TestCheckpointTimeout(c *C) {
 		c.Assert(err, IsNil)
 	})
 	log.Info("Completed test TestCheckpointTimeout")
+}
+
+func (s *IntegSuite) TestSavepointCheckpointFailureFallback(c *C) {
+	log.Info("Starting test TestSavepointCheckpointFailureFallback")
+	// create a Flink app
+	testName := "recoveryfallback"
+	config, err := s.Util.ReadFlinkApplication("test_app.yaml")
+	c.Assert(err, IsNil, Commentf("Failed to read test app yaml"))
+	config.Name = testName + "job"
+	config.Spec.DeleteMode = v1beta1.DeleteModeForceCancel
+	config.Spec.FallbackWithoutState = true
+	config.ObjectMeta.Labels["integTest"] = testName
+
+	c.Assert(s.Util.CreateFlinkApplication(config), IsNil,
+		Commentf("Failed to create flink application"))
+
+	c.Assert(s.Util.WaitForPhase(config.Name, v1beta1.FlinkApplicationRunning, v1beta1.FlinkApplicationDeployFailed), IsNil)
+
+	// Cause it to fail
+	err = s.Util.ExecuteCommand("minikube", "ssh", "touch /tmp/checkpoints/fail && chmod 0644 /tmp/checkpoints/fail")
+	c.Assert(err, IsNil)
+
+	// wait a bit for it to start failing
+	time.Sleep(5 * time.Second)
+
+	// get app details
+	app, err := s.Util.GetFlinkApplication(config.Name)
+	c.Assert(err, IsNil)
+
+	// Try to update it with app that does not fail on checkpoint
+	newApp := WaitUpdateAndValidate(c, s, config.Name, func(app *v1beta1.FlinkApplication) {
+		skipFailureEnvVar := v12.EnvVar{Name: "SKIP_INDUCED_FAILURE", Value: "true"}
+		app.Spec.Image = NewImage
+		app.Spec.JobManagerConfig.EnvConfig.Env = append(app.Spec.JobManagerConfig.EnvConfig.Env, skipFailureEnvVar)
+		app.Spec.TaskManagerConfig.EnvConfig.Env = append(app.Spec.TaskManagerConfig.EnvConfig.Env, skipFailureEnvVar)
+	}, v1beta1.FlinkApplicationDeployFailed)
+
+	// Check job updated and started without savepointPath
+	c.Assert(newApp.Status.JobStatus.JobID, Not(Equals), app.Status.JobStatus.JobID)
+	c.Assert(newApp.Spec.SavepointPath, Equals, "")
+
+	// Check new app has no failures
+	endpoint := fmt.Sprintf("jobs/%s", newApp.Status.JobStatus.JobID)
+	_, err = s.Util.FlinkAPIGet(newApp, endpoint)
+	c.Assert(err, IsNil)
+
+	// delete the application and ensure everything is cleaned up successfully
+	c.Assert(s.Util.FlinkApps().Delete(newApp.Name, &v1.DeleteOptions{}), IsNil)
+
+	for {
+		pods, err := s.Util.KubeClient.CoreV1().Pods(s.Util.Namespace.Name).
+			List(v1.ListOptions{LabelSelector: "integTest=" + testName})
+		c.Assert(err, IsNil)
+		if len(pods.Items) == 0 {
+			break
+		}
+	}
+	log.Info("All pods torn down")
+	log.Info("Completed test TestJobCancellationWithoutSavepoint")
 }
