+++ integ/job_cancellation_test.go
@@ -59,7 +59,6 @@ func WaitUpdateAndValidate(c *C, s *IntegSuite, name string, updateFn func(app *
 // tests the workflow of job cancellation without savepoint
 func (s *IntegSuite) TestJobCancellationWithoutSavepoint(c *C) {
 	log.Info("Starting test TestJobCancellationWithoutSavepoint")
-	c.Skip("local")
 	testName := "cancelsuccess"
 	const finalizer = "simple.finalizers.test.com"
 
@@ -139,7 +138,7 @@ func (s *IntegSuite) TestJobCancellationWithoutSavepoint(c *C) {
 // here, the new submitted job starts without a savepoint.
 func (s *IntegSuite) TestCancelledJobWithoutSavepoint(c *C) {
 	log.Info("Starting test TestCancelledJobWithoutSavepoint")
-	c.Skip("local")
+
 	testName := "invalidcancel"
 	config, err := s.Util.ReadFlinkApplication("test_app.yaml")
 	c.Assert(err, IsNil, Commentf("Failed to read test app yaml"))
@@ -215,7 +214,6 @@ func (s *IntegSuite) TestCancelledJobWithoutSavepoint(c *C) {
 // tests the recovery workflow of the job when savepoint is disabled.
 func (s *IntegSuite) TestJobRecoveryWithoutSavepoint(c *C) {
 	log.Info("Starting test TestJobRecoveryWithoutSavepoint")
-	c.Skip("local")
 
 	const finalizer = "simple.finalizers.test.com"
 	const testName = "cancelrecovery"
