+++ integ/simple_test.go
@@ -15,7 +15,7 @@ import (
 )
 
 // const NewImage = "lyft/operator-test-app:b1b3cb8e8f98bd41f44f9c89f8462ce255e0d13f.2"
-const NewImage = "operator-test-app:test11"
+const NewImage = "operator-test-app:local.2"
 
 func updateAndValidate(c *C, s *IntegSuite, name string, updateFn func(app *v1beta1.FlinkApplication), failurePhase v1beta1.FlinkApplicationPhase) *v1beta1.FlinkApplication {
 	app, err := s.Util.Update(name, updateFn)
@@ -70,7 +70,6 @@ func updateAndValidate(c *C, s *IntegSuite, name string, updateFn func(app *v1be
 // Tests job submission, upgrade, rollback, and deletion
 func (s *IntegSuite) TestSimple(c *C) {
 	log.Info("Starting test TestSimple")
-	c.Skip("local")
 
 	const finalizer = "simple.finalizers.test.com"
 
@@ -269,7 +268,6 @@ func (s *IntegSuite) TestSimple(c *C) {
 
 func (s *IntegSuite) TestRecovery(c *C) {
 	log.Info("Starting test TestRecovery")
-	c.Skip("local")
 
 	config, err := s.Util.ReadFlinkApplication("test_app.yaml")
 	c.Assert(err, IsNil, Commentf("Failed to read test app yaml"))
