+++ integ/test_app.yaml
@@ -6,7 +6,7 @@ metadata:
   labels:
     environment: development
 spec:
-  image: operator-test-app:test11b
+  image: operator-test-app:local.1
   imagePullPolicy: IfNotPresent
   imagePullSecrets:
     - name: dockerhub
